server: move hot region scheduler into schedulers directory. (#730)
